src/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
google-cloud-secret-manager==2.18.0
google-cloud-storage==2.14.0
google-auth==2.25.2
